Split controllers out into individual files and use gulp to combine
Generates the application.js from component parts. A bit messy but
easier to wrangle the source files - a poor simulation of how webpack
does it, but compiling exported modules with webpack doesn't work when
"serving" the HTML file from the local filesystem.

Sadly, accessing the sourcemap from the local filesystem doesn't work
either but I'm going to leave the line there anyway
