+++ .gitignore
@@ -1,5 +1,6 @@
 index.html
 assets/style.css
+assets/script/application*
 .DS_Store
 .sass-cache
 app/test/assets/**/*
