+++ app/src/app_start.js
@@ -1,0 +1,2 @@
+(() => {
+  const application = Stimulus.Application.start()
