+++ app/src/app_stop.js
@@ -1,0 +1,2 @@
+})()
+//# sourceMappingURL=localhost://assets/script/application.js.map
