+++ app/src/controllers/notbook_controller.js
@@ -1,0 +1,8 @@
+  application.register("notbook", class extends Stimulus.Controller {
+    close_books() {
+      document.querySelectorAll('.about').forEach((element, idx) => {
+        element.style.display = "none"
+      })
+      document.querySelector('#fade').style.display = "none"
+    }
+  })
